Move augmented_sys_path under tests
